app/requirements.txt
# app/requirements.txt
google-adk
google-cloud-discoveryengine
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
uvicorn
pytest
